feat: add spell slot display component
