+++ client/src/App.scss
@@ -39,6 +39,56 @@
   text-align: center;
 }
 
+.spell-slot-container {
+  position: fixed;
+  bottom: 60px;
+  left: 0;
+  right: 0;
+  display: flex;
+  justify-content: center;
+  gap: 0.5rem;
+  z-index: 1030;
+}
+
+.spell-slot {
+  width: 48px;
+  height: 48px;
+  border: 2px solid #fff;
+  background: rgba(0, 0, 0, 0.7);
+  color: #fff;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: space-between;
+  padding: 2px;
+}
+
+.spell-slot .slot-level {
+  font-size: 0.75rem;
+}
+
+.spell-slot .slot-boxes {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2px;
+  justify-content: center;
+}
+
+.spell-slot .slot-small {
+  width: 10px;
+  height: 10px;
+  border: 1px solid #fff;
+  cursor: pointer;
+}
+
+.spell-slot .slot-filled {
+  background: #fff;
+}
+
+.spell-slot .slot-used {
+  background: transparent;
+}
+
 .text-center {
   text-align: center;
 }
