+++ server/data/classFeatures.js
@@ -22,75 +22,12 @@ const ASI = {
     'Increase one ability score of your choice by 2, or increase two ability scores by 1. You can’t increase an ability score above 20.'
 };
 
-// Spell slot progression tables
-const fullCasterSlots = {
-  1: { 1: 2 },
-  2: { 1: 3 },
-  3: { 1: 4, 2: 2 },
-  4: { 1: 4, 2: 3 },
-  5: { 1: 4, 2: 3, 3: 2 },
-  6: { 1: 4, 2: 3, 3: 3 },
-  7: { 1: 4, 2: 3, 3: 3, 4: 1 },
-  8: { 1: 4, 2: 3, 3: 3, 4: 2 },
-  9: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
- 10: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
- 11: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
- 12: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1 },
- 13: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
- 14: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1 },
- 15: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
- 16: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1 },
- 17: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2, 6: 1, 7: 1, 8: 1, 9: 1 },
- 18: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 1, 7: 1, 8: 1, 9: 1 },
- 19: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 1, 8: 1, 9: 1 },
- 20: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 3, 6: 2, 7: 2, 8: 1, 9: 1 }
-};
-
-const halfCasterSlots = {
-  1: {},
-  2: { 1: 2 },
-  3: { 1: 3 },
-  4: { 1: 3 },
-  5: { 1: 4, 2: 2 },
-  6: { 1: 4, 2: 2 },
-  7: { 1: 4, 2: 3 },
-  8: { 1: 4, 2: 3 },
-  9: { 1: 4, 2: 3, 3: 2 },
- 10: { 1: 4, 2: 3, 3: 2 },
- 11: { 1: 4, 2: 3, 3: 3 },
- 12: { 1: 4, 2: 3, 3: 3 },
- 13: { 1: 4, 2: 3, 3: 3, 4: 1 },
- 14: { 1: 4, 2: 3, 3: 3, 4: 1 },
- 15: { 1: 4, 2: 3, 3: 3, 4: 2 },
- 16: { 1: 4, 2: 3, 3: 3, 4: 2 },
- 17: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
- 18: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 1 },
- 19: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 },
- 20: { 1: 4, 2: 3, 3: 3, 4: 3, 5: 2 }
-};
-
-const pactMagic = {
-  1: { 1: 1 },
-  2: { 1: 2 },
-  3: { 2: 2 },
-  4: { 2: 2 },
-  5: { 3: 2 },
-  6: { 3: 2 },
-  7: { 4: 2 },
-  8: { 4: 2 },
-  9: { 5: 2 },
- 10: { 5: 2 },
- 11: { 5: 3 },
- 12: { 5: 3 },
- 13: { 5: 3 },
- 14: { 5: 3 },
- 15: { 5: 3 },
- 16: { 5: 3 },
- 17: { 5: 4 },
- 18: { 5: 4 },
- 19: { 5: 4 },
-  20: { 5: 4 }
-};
+// Spell slot progression tables shared with the client
+const {
+  fullCasterSlots,
+  halfCasterSlots,
+  pactMagic,
+} = require('../../client/src/utils/spellSlots');
 
 // Spells known tables
 function paladinSpellsKnown(level, abilityMod) {
